+++ src/main/java/seedu/taskmaster/model/ModelManager.java
@@ -174,19 +174,19 @@ public class ModelManager implements Model {
     }
 
     @Override
-    public void scoreStudent(StudentRecord target, int score) {
+    public void scoreStudent(StudentRecord target, double score) {
         requireAllNonNull(target, score);
         taskmaster.scoreStudent(target, score);
     }
 
     @Override
-    public void scoreStudentWithNusnetId(NusnetId nusnetId, int score) {
+    public void scoreStudentWithNusnetId(NusnetId nusnetId, double score) {
         requireAllNonNull(nusnetId, score);
         taskmaster.scoreStudentWithNusnetId(nusnetId, score);
     }
 
     @Override
-    public void scoreAllStudents(List<StudentRecord> students, int score) {
+    public void scoreAllStudents(List<StudentRecord> students, double score) {
         List<NusnetId> nusnetIds = students
                 .stream()
                 .map(StudentRecord::getNusnetId)
